scribe: restructure and update
