+++ packages/scribe/src/components/FloatingBox/FloatingBox.tsx
