+++ packages/scribe/src/components/NodesMenu/Menu/MenuContext.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react-refresh/only-export-components */
 import { createContext, useContext } from "react";
 import { useMenuCore } from "./useMenuCore";
