+++ packages/scribe/src/components/NodesMenu/Menu/types.d.ts
