+++ packages/scribe/src/editor/NoteEditor.tsx
@@ -1,6 +1,6 @@
-import editorUsjAdaptor from "../adaptors/editor-usj.adaptor";
-import usjNoteEditorAdapter from "../adaptors/note-usj-editor.adaptor";
-import editorTheme from "../themes/editor-theme";
+import editorUsjAdaptor from "./adaptors/editor-usj.adaptor";
+import usjNoteEditorAdapter from "./adaptors/note-usj-editor.adaptor";
+import editorTheme from "./themes/editor-theme";
 import { Usj } from "@biblionexus-foundation/scripture-utilities";
 import { AutoFocusPlugin } from "@lexical/react/LexicalAutoFocusPlugin";
 import { LexicalComposer } from "@lexical/react/LexicalComposer";
