+++ packages/scribe/src/editor/adaptors/note-usj-editor.adaptor.ts
