+++ packages/scribe/src/editor/nodes/emptyFootNote.ts
