+++ packages/scribe/src/editor/plugins/KeyboardShortcutPlugin.tsx
