+++ packages/scribe/src/editor/plugins/ScriptureReferencePlugin.tsx
