+++ packages/scribe/src/editor/plugins/view-mode.model.ts
