+++ packages/scribe/src/editor/themes/editor-theme.ts
