+++ packages/scribe/src/editor/utils/insertFunctions.ts
