+++ packages/scribe/src/hooks/useCursorCoords.tsx
