+++ packages/scribe/src/hooks/usePointerInteractions.tsx
