+++ packages/scribe/src/index.ts
@@ -1,4 +1,4 @@
-export { default as Editor, type EditorRef, type EditorOptions } from "./components/Editor";
+export { default as Editor, type EditorRef, type EditorOptions } from "./editor/Editor";
 export { immutableNoteCallerNodeName } from "shared-react/nodes/usj/ImmutableNoteCallerNode";
 export { type UsjNodeOptions } from "shared-react/nodes/usj/usj-node-options.model";
 export { type ViewMode } from "shared-react/views/view-mode.model";
