+++ packages/scribe/src/styles/index.css
